// desktop/src/types.ts
type TMaybe<T> = T | null | undefined
// ...
export type TProviderID = string
// ...
export type TProviderOptions = Record<string, TProviderOption>
export type TProviderOption = Readonly<{
  // Value is the options current value
  value: TMaybe<string>
  // Children are the children generated by this option
  children: TMaybe<string[]>
  // If value is a password
  password: TMaybe<boolean>
  // DisplayName of the option, preferred over the option name by a supporting tool.
  displayName: TMaybe<string>
  // A description of the option displayed to the user by a supporting tool.
  description: TMaybe<string>
  // If required is true and the user doesn't supply a value, devsy will ask the user
  required: TMaybe<boolean>
  // Allowed values for this option.
  enum: TMaybe<TOptionEnum[]>
  // Suggestions are suggestions to show in the Devsy UI for this option
  suggestions: TMaybe<string[]>
  // Hidden specifies if the option should be hidden
  hidden: TMaybe<boolean>
  // Local means the variable is not resolved immediately and instead later when the workspace / machine was created.
  local: TMaybe<boolean>
  // Default value if the user omits this option from their configuration.
  default: TMaybe<string>
  // Command is the command to run to specify an option
  command: TMaybe<string>
  // SubOptionsCommand is the command to retrieve sub options
  subOptionsCommand: TMaybe<string>
  // Type is the provider option type. Can be one of: string, multiline, duration, number or boolean. Defaults to string
  type: TMaybe<"string" | "duration" | "number" | "boolean" | "multiline">
  // Mutable specifies if an option can be changed on the workspace or machine after creating it
  mutable: TMaybe<boolean>
}>
// ...
export type TWorkspaceID = NonNullable<TWorkspace["id"]>
export type TWithWorkspaceID = Readonly<{ workspaceID: TWorkspaceID }>
export type TWorkspace = Readonly<{
  id: string
  uid: string
  picture: TMaybe<string>
  machine: TMaybe<Readonly<{ machineId: TMaybe<string> }>>
  provider: TMaybe<Readonly<{ name: TMaybe<string>; options: TMaybe<TProviderOptions> }>>
  status: TMaybe<"Running" | "Busy" | "Stopped" | "NotFound">
  ide: TMaybe<{
    name: TMaybe<string>
  }>
  creationTimestamp: string
  lastUsed: string
  source: TMaybe<TWorkspaceSource>
}>
export type TWorkspaceWithoutStatus = Omit<TWorkspace, "status"> & Readonly<{ status: null }>
export type TWorkspaceStatusResult = Readonly<{
  id: TMaybe<string>
  context: TMaybe<string>
  provider: TMaybe<string>
  state: TMaybe<TWorkspace["status"]>
}>
export type TWorkspaceSourceType = "local" | "git" | "image"
export type TWorkspaceStartConfig = Readonly<{
  id: string
  prebuildRepositories?: string[]
  devcontainerPath?: string
  ideConfig?: TWorkspace["ide"]
  providerConfig?: Readonly<{ providerID?: TProviderID; options?: Record<string, string> }>
  // Instead of starting a workspace just by ID, the sourceConfig starts it with a `source/ID` combination
  sourceConfig?: Readonly<{
    source: string
    type?: TWorkspaceSourceType
  }>
}>
export type TWorkspaceSource = {
  gitRepository: TMaybe<string>
  gitBranch: TMaybe<string>
  gitCommit: TMaybe<string>
  gitPRReference: TMaybe<string>
  gitSubPath: TMaybe<string>
  localFolder: TMaybe<string>
  image: TMaybe<string>
}